+++ src/api/images/handlers/delete.ts
@@ -6,6 +6,11 @@ import prisma from "../../../../prisma/client";
 const deleteImage: ImageHandlers["delete"] = async (req, res) => {
   const { id } = req.params;
   try {
+    if (!req.user || req.user.role !== "SUPER_ADMIN") {
+      return res
+        .status(403)
+        .json({ message: "Forbidden, you don't have the right access" });
+    }
     const deleteImage = await prisma.image.delete({
       where: {
         id: id,
